Polygon

+++ model/join/orders_detail.go
@@ -24,7 +24,7 @@ type OrdersDetail struct {
 	IndexID     uint64    `gorm:"column:index_id" json:"index_id"`
 	NickName    string    `gorm:"column:nick_name" json:"nick_name"`
 	Description string    `gorm:"column:description" json:"description"`
-	OriginChain uint8     `gorm:"column:origin_chain" json:"origin_chain"`
+	OriginChain uint64    `gorm:"column:origin_chain" json:"origin_chain"`
 	GroupName   string    `gorm:"column:group_name" json:"group_name"`
 }
